requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
blobfile==3.0.0
cachetools==5.3.2
catboost==1.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
diffusers==0.30.3
einops==0.8.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.26.0
idna==3.6
importlib-metadata==8.5.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jupyter-client==8.6.0
jupyter-core==5.5.0
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
lxml==5.3.0
markdown==3.5.1
markupsafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.21.0
pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.2
pyzmq==25.1.2
regex==2024.9.11
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
spyder-kernels==2.5.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.6.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
triton==2.1.0
typing-extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
werkzeug==3.0.1
wrapt==1.14.1
wurlitzer==3.0.3
xgboost==2.0.2
zipp==3.20.2
